Struct google-freebase1::ReconcileGet

pub struct ReconcileGet {
    pub costs: ReconcileGetCosts,
    pub warning: Vec<ReconcileGetWarning>,
    pub candidate: Vec<ReconcileCandidate>,
    pub match_: ReconcileCandidate,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

reconcile (response)

Fields

costs

Server costs for reconciling.

warning

If filled, then there were recoverable problems that affected the request. For example, some of the properties were ignored because they either are not valid Freebase predicates or are not indexed for reconciliation. The candidates returned should be considered valid results, with the caveat that sections of the request were ignored as specified by the warning text.

candidate

If filled, then the listed candidates are potential matches, and such should be evaluated by a more discerning algorithm or human. The matches are ordered by confidence.

match_

If filled, this entity is guaranteed to match at requested confidence probability (default 99%).